pyproject.toml
[tool.poetry]
name = "img-face-cluster"
version = "0.1.0"
description = "Clustering faces found in photos"
authors = ["Jakub Ciszek <jakubciszek64@gmail.com>"]
license  = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
torch = "^1.9.0"
torchvision = "^0.10.0"
scikit-learn = "^0.24.2"
facenet-pytorch = "^2.5.2"
Pillow = "^8.3.1"
pandas = "^1.3.1"
matplotlib = "^3.4.2"
seaborn = "^0.11.2"
SQLAlchemy = "^1.4.22"
tqdm = "^4.62.1"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
ipykernel = "^6.0.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
